refactor: change id type of entity and dto to long, reuse collectData() method in createEvent() of EventController class, delete unused readByte() method in ConsoleReader class

+++ src/main/java/org/agenda/event/dto/EventResponse.java
@@ -3,5 +3,5 @@ package org.agenda.event.dto;
 import java.time.LocalDateTime;
 import java.util.List;
 
-public record EventResponse (int id, String title, LocalDateTime date, List<String> warnings){
+public record EventResponse (long id, String title, LocalDateTime date, List<String> warnings){
 }
