+++ src/main/java/org/agenda/event/repository/EventNotFoundException.java
@@ -1,7 +1,7 @@
 package org.agenda.event.repository;
 
 public class EventNotFoundException extends RuntimeException {
-    public EventNotFoundException(String failedPerformance, int id) {
+    public EventNotFoundException(String failedPerformance, long id) {
         super(String.format("Failed %s: Event with id %s not found", failedPerformance, id));
     }
 }
